FINAL VERSION FOR SUBMISSION
Finalized styling for all pages

added and completed a timeline page

Fixed the display issues with the radar graph on the comparison page

Ensured Plane BR changes when ground battle is ticked

All filter logic works now for all nations

search feature is more refined and doesnt require us to be too specific

Timeline graph works how ever the versions of the patches we are pulling are very limited and and changes are found in a very select vehicles but differences can be seen.

Fixed the chart on the landing page and made sure its a custom one as per brief

Responsive design was attempted but not very succesful

All pages are connected and setup up corretly in the NAV

Played around with colours for beter clarity

CHATGPT was used to help figure out my timeline issues because the Schemas from the documentation wasnt very helpful and ended using a file found on the API github page to better understand the call functions with the timeline feature.

GLOBAL CSS page was later made with the help of chatgpt to make sure everything was neat and working.

Tweaks were made to the API to make for better handling of data , made use of pagnitation so we could load all data and not just the default 200 limit as well to reduce latency